# packages/contracts/src/generated/model/categoryInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TaskNest API
 * TaskNest REST API. Every endpoint except /healthz requires a Clerk session
 * (`Authorization: Bearer <session token>`). All data is private to the signed-in user.
 *
 * Clients may send `X-Timezone: <IANA zone>` (e.g. `Europe/Athens`); "today" in task
 * filters and summaries is then computed in that zone. Defaults to UTC.
 *
 * OpenAPI spec version: 1.0.0
 */
import type { HexColor } from './hexColor';

export interface CategoryInput {
  /**
     * @minLength 1
     * @maxLength 60
     */
  name: string;
  color?: HexColor;
}